Differentiate between account users and other users on the same server.
Two different users may be on the same server. For that reason we have
to be more precise what user we want to target. For example, to send a
message to a user we have to check if we allowed to post to this user
and we have to find the correct mailbox.